Unit tests for error handling in invocation index list component.
